Move logic to other file
